Add secondary bot setup to only make findTrollHouses request
This is important because if the targer trollHouse banned this bot, it
will fail to find troll with this error:

    {
        "description": "Forbidden: bot was kicked from the supergroup chat",
        "error_code": 403,
        "ok": false
    }

However, if we use a hidden bot (unknown from trolls) to only make
this request, it works pretty fine. After all, they never will know
which bot to ban to stop the troll-shield.