Add support to qualified command like /pass@lelerax_bot (#11)
* Add support to qualified command like /pass@lelerax_bot

I finally registered the bot commands via API, so we have autocomplete
on Telegram by using the bot. In groups using /pass or /kills would be
translated to /pass@lelerax_bot and /kills@lelerax_bot, but in v0.7.0
troll-shield version this is not supported.

This PR takes care to support this properly, giving attention to edge
cases like using /pass@anotherBotName (this will not work). /pass
/kill simplified syntax will always works, but can trigger multiple
bots if they implements the same command.

CHANGELOG:
- Add checkCommand which checks properly if is a valid command
- Add commandEvent which can be use to filter commands (performance)
- Add unit tests for both functions
- Rename /lelerax to /ping, which makes more sense.
- Add renan_r as admin.

* Remove funlen linter, too much annoying